+++ src/smk-settings.adb
@@ -29,9 +29,19 @@ with Ada.Strings.Unbounded; use Ada.Strings.Unbounded;
 
 package body Smk.Settings is
 
-   Smkfl_Name  : Unbounded_String := Null_Unbounded_String;
-   Runfl_Name  : Unbounded_String := Null_Unbounded_String;
-   Target_Name : Unbounded_String := Null_Unbounded_String;
+   Smkfl_Name      : Unbounded_String := Null_Unbounded_String;
+   Runfl_Name      : Unbounded_String := Null_Unbounded_String;
+   Current_Section : Unbounded_String := Null_Unbounded_String;
+   Cmd_Line        : Unbounded_String := Null_Unbounded_String;
+   Current_Target  : Unbounded_String := Null_Unbounded_String;
+
+   -- --------------------------------------------------------------------------
+   function Run_Dir_Name return String is
+     (Ada.Directories.Containing_Directory (Smkfile_Name));
+
+   function Strace_Outfile_Name return String is
+     (Strace_Outfile_Prefix & Ada.Directories.Simple_Name (Smkfile_Name)
+      & Strace_Outfile_Suffix);
 
    -- --------------------------------------------------------------------------
    procedure Set_Smkfile_Name (Name : in String) is
@@ -40,31 +50,47 @@ package body Smk.Settings is
       Runfl_Name := To_Unbounded_String (To_Runfile_Name (Name));
    end Set_Smkfile_Name;
 
+   function Smkfile_Name return String is (To_String (Smkfl_Name));
+
    -- --------------------------------------------------------------------------
    procedure Set_Runfile_Name (Name : in String) is
    begin
       Runfl_Name := To_Unbounded_String (Name);
    end Set_Runfile_Name;
 
-   -- --------------------------------------------------------------------------
-   procedure Set_Target (Name : in String) is
-   begin
-      Target_Name := To_Unbounded_String (Name);
-   end Set_Target;
+   function Runfile_Name return String is (To_String (Runfl_Name));
 
    -- --------------------------------------------------------------------------
    function To_Runfile_Name (Smkfile_Name : in String) return String is
-      (Smk_File_Prefix & Ada.Directories.Simple_Name (Smkfile_Name));
+     (Smk_File_Prefix & Ada.Directories.Simple_Name (Smkfile_Name));
 
    -- --------------------------------------------------------------------------
-   function Smkfile_Name return String is (To_String (Smkfl_Name));
-   function Runfile_Name return String is (To_String (Runfl_Name));
-   function Run_Dir_Name return String is
-     (Ada.Directories.Containing_Directory (Smkfile_Name));
-   function Strace_Outfile_Name return String is
-     (Strace_Outfile_Prefix & Ada.Directories.Simple_Name (Smkfile_Name)
-      & Strace_Outfile_Suffix);
-   function Target return String is (To_String (Target_Name));
+   procedure Set_Section_Name (Name : in String) is
+   begin
+      Current_Section := To_Unbounded_String (Name);
+   end Set_Section_Name;
+
+   function Section_Name return String is (To_String (Current_Section));
+
+   -- --------------------------------------------------------------------------
+   procedure Add_To_Command_Line (Text : in String) is
+   begin
+      if Cmd_Line = "" then
+         Cmd_Line := To_Unbounded_String (Text);
+      else
+         Cmd_Line := Cmd_Line & " " & Text;
+      end if;
+   end Add_To_Command_Line;
+
+   function Command_Line return String is (To_String (Cmd_Line));
+
+   -- --------------------------------------------------------------------------
+   procedure Set_Target_Name (Target : in String) is
+   begin
+      Current_Target := To_Unbounded_String (Target);
+   end Set_Target_Name;
+
+   function Target_Name return String is (To_String (Current_Target));
 
    -- --------------------------------------------------------------------------
    function Is_System_File (File_Name : in String) return Boolean is